ShannonDiversity Shannon Diversity Metrics

Calculates a number of metrics based on the Shannon information entropy measure of diversity in a vector, x.

Usage

ShannonDiversity(x)

Arguments

x

1 x n vector.

Value

H

Shannon entropy-based metric of diversity. This captures the effects of both richnes (the length of the vector, n) and the evenennes of the distribution.

Hmax

The maximum possible value of H given a vector of the length n provided.

Hr

Relative evenness Hr = H/Hmax

Hcentral

The centralization or concentration of the values among the n elements

n

Number of elements in the vector.

effective.n

effective number of elements in the vector, given the distribution of the relative weights.

Note

The formulation for Shannon Diversity uses a natural logarithm. As the natural logarithm of zero is undefined, the input vector cannot contain zeros. Analytically, there are two approaches to dealing with this issue if your vector contains zeros. First, you can apply the analysis to only the non-zero elements. Second, you can add a tiny amount to all of the elements such that the zero elements are now very small numbers, relative the original vector values.

Author

Stuart R. Borrett

Examples

data(oyster)

## throughflow diversity
T <- enaFlow(oyster)$T
ShannonDiversity(T)
#>           H        Hmax          Hr    Hcentral           n effective.n 
#>   1.3042705   1.7917595   0.7279272   0.2720728   6.0000000   3.6849998 

## storage (biomass) biodiversity
## X <- oyster %v% "storage"
## ShannonDiversity(X)