Calculate the Total Environ Throughflow

Determines the total environ throughflow (TET) for each of the 2 x n environs of the selected network model. It returns both the TET calculated from a unit input (output) vector and from the observed or realized input (output) vector.

Usage

TET(x, balance.override = FALSE)

Arguments

x

A network object.

balance.override

Logical: should the function work if the model is not at steady-state?

Value

realized.input

vector of the n realized total environ throughflows for the n input oriented environs.

realzied.output

vector of the n realized total environ throughflows for the n ouptut oriented environs.

unit.input

vector of the n unit total environ throughflows for the n input oriented environs.

unit.output

vector of the n unit total environ throughflows for the n output oriented environs.

References

Gattie, D.K., Schramski, J.R., Borrett, S.R., Patten, B.C., Bata, S.A., and Whipple, S.J. 2006. Indirect effects and distributed control in ecosystems: Network environ analysis of a seven-compartment model of nitrogen flow in the Neuse River Estuary, USA—Steady-state analysis. Ecol. Model. 194:162–177.

Whipple, S.J., Borrett, S.R., Patten, B.C., Gattie, D.K., Schramski, J.R., and Bata, S.A. 2007. Indirect effects and distributed control in ecosystems: Comparative network environ analysis of a seven-compartment model of nitrogen flow in the Neuse River Estuary, USA—Time series analysis. Ecol. Model. 206: 1–17.

See also

Author

Matthew K. Lau Stuart R. Borrett

Examples

data(troModels)
tet <- TET(troModels[[6]])
#> Warning: NOTE: output of functions from a particular analytical school will be returned in the standard orientation of that school.
tet
#> $realized.input
#> [1]  2303.000 11880.010  6161.980   892.571  9388.439
#> 
#> $realized.output
#> [1] 29328.641     0.000     0.000     0.000  1297.359
#> 
#> $unit.input
#> [1] 1.000000 3.365442 3.396902 4.396902 2.365442
#> 
#> $unit.output
#> [1] 2.622375 1.649215 1.469720 1.922149 2.043085
#>