Calculates the Ascendency of an Ecological Network

Calculates the average mutual information (AMI), ascendency, overhead, and capacity of input-output networks. It also returns the ratios of ascendency and overhead to capacity. These metrics describe the organization of flow in an ecological network (Ulanowicz 1997).

Usage

enaAscendency(x = "network object")

Arguments

x

A network object.

Value

H

Total flow diversity (MacArthur 1955). Uses the Shannon Information measure (aka Boltzmann entropy) applied to the individual flows.

AMI

Returns the Average Mutual Information (AMI) in a network. AMI provides a measure of the constraints placed on a given peice of energy matter moving through a network (Patricio et al. 2006)

Hr

Residual uncertainty that remains about the flow distribution once the ecosystem structure is specified (Hr = H - AMI).

ASC

Returns the ascendnecy of a network. Ascendency is a scaled form of AMI relative to the total system throughput (Ulanowicz 1997; 2004). Total system throughput is the sum of all activity in a network (Kay et al. 1989).

OH

Returns the overhead of a network. Overhead is the proportion of the capacity in a network that is not used as ascendency (Ulanowicz 2004).

CAP

Returns the capacity of a network. Capacity is defined as the sum of ascendency and overhead (Ulanowicz 2004).

ACS.CAP

Returns the proportion of capacity used by ascendency.

OH.CAP

Returns the proportion of capacity used by overhead.

robustness

Returns the robustness of the network.

ELD

Returns the Effective Link Density of the network(c) (Ulanowicz et al. 2014).

TD

Returns the Trophic Depth of the network(r) (Ulanowicz et al. 2014).

A.input

Returns the input ascendnecy of a network.

A.internal

Returns the internal ascendnecy of a network.

A.export

Returns the export ascendnecy of a network.

A.respiration

Returns the respiration ascendnecy of a network.

OH.input

Returns the input overhead of a network.

OH.internal

Returns the internal overhead of a network.

OH.export

Returns the export overhead of a network.

OH.respiration

Returns the respiration overhead of a network.

CAP.input

Returns the input capacity of a network.

CAP.internal

Returns the internal capacity of a network.

CAP.export

Returns the export capacity of a network.

CAP.respiration

Returns the respiration capacity of a network.

Details

This and other Ulanowicz school functions require that export and respiration components of output be separately quantified.

References

Kay, J.J., Graham, L.A., Ulanowicz, R.E., 1989. A detailed guide to network analysis. p. 15-61 In: Wulff, F., Field, J.G., Man, K.H. (eds.) Network analysis in marine ecology. Coastal Estuarine Study Serries. Springer-Verlag, Berlin.

Patrico, J., Ulanowicz, R.E., Pardal, M.A., Marques J.C., 2004. Ascendency as an ecological indicator: a case study of estuarine pulse eutrophication. Estuar. Coast Shelf S. 60, 23-35.

Ulanowicz, R.E. and Norden, J.S., 1990. Symmetrical overhead in flow networks. International Journal of Systems Science, 21(2), pp.429-437.

Ulanowicz, R.E., 1997. Ecology, The Ascendent Perspective. Columbia University Press, New York.

Ulanowicz, R.E., 2004. Quantitative methods for ecological network analysis. Comput. Biol. Chem. 28, 321-33

Ulanowicz, R.E., Holt, R.D., Barfield, M., 2014. Limits on ecosystem trophic complexity: insights from ecological network analysis. Ecology Letters 17:127-136

See also

Author

David E. Hines Matthew K. Lau Stuart R. Borrett

Examples

data(troModels)
enaAscendency(troModels[[6]])
#>            H      AMI       Hr      CAP      ASC       OH   ASC.CAP    OH.CAP
#> [1,] 3.20096 1.336447 1.864513 135864.7 56725.49 79139.25 0.4175144 0.5824856
#>      robustness      ELD       TD  A.input A.internal A.export A.respiration
#> [1,]  0.3646722 1.908258 2.525286 19147.85   29331.98 1051.765      7193.894
#>      OH.input OH.internal OH.export OH.respiration CAP.input CAP.internal
#> [1,] 6221.751    29832.46  7810.812       35274.22   25369.6     59164.44
#>      CAP.export CAP.respiration
#> [1,]   8862.578        42468.11