Control Analyses of Ecological Networks

Analyses for analyzing the control amongst the nodes in ecological networks.

Usage

enaControl(x, zero.na = TRUE, balance.override = FALSE)

Arguments

x

A network object.

zero.na

Makes undefined (NA) values zero.

balance.override

Turns off balancing and checks of network balance.

Value

CN

Control matrix using flow values.

CQ

Control matrix using storage values.

CR

Schramski Control Ratio Matrix

CD

Schramski Control Difference Matrix

CA

Control Allocation Matrix

CDep

Control Dependency Matrix

sc

Schramski System Control vector

scp

Schramski system control vector as percent of total control

ns

vector of network-level summary statistics

References

Fath, B. D., Borrett, S. R. 2006. A MATLAB function for Network Environ Analysis. Environmental Modelling & Software 21:375-405

Schramski, J.R., Gattie, D.K., Patten, B.C., Borrett S.R., Fath, B.D., Thomas, C.R., and Whipple, S.J. 2006. Indirect effects and distributed control in ecosystems: Distributed control in the environ networks of a seven compartment model of nitrogen flow in the Neuse River Estuary, USA Steady-state analysis. Ecological Modelling 194:189-201

Schramski, J.R., Gattie, D.K., Patten, B.C., Borrett S.R., Fath, B.D., and Whipple, S.J. 2007. Indirect effects and distributed control in ecosystems: Distributed control in the environ networks of a seven compartment model of nitrogen flow in the Neuse River Estuary, USA Time series analysis. Ecological Modelling 206:18-30

Chen, S., Fath, B.D., Chen, B. 2011. Information-based network environ analysis: a system perspective for ecologcial risk assessment. Ecol. Ind. 11:1664-1672.

Chen, S. and Chen, B. 2015. Urban energy consumption: Different insights from energy flow analysis, input-output analysis and ecological network analysis. Applied Energy 138:99-107.

See also

Author

Matthew K. Lau Stuart R. Borrett Pawandeep Singh

Examples

data(troModels)
enaControl(troModels[[6]])
#> $CN
#>                  PLANTS  BACTERIA DETRITUS FEEDERS CARNIVORES  DETRITUS
#> PLANTS                0 0.0000000         0.000000  0.0000000 0.0000000
#> BACTERIA              1 0.0000000         0.000000  0.1865961 0.6903851
#> DETRITUS FEEDERS      1 0.5806865         0.000000  0.0000000 0.8460570
#> CARNIVORES            1 0.0000000         0.891124  0.0000000 0.5486486
#> DETRITUS              1 0.0000000         0.000000  0.0000000 0.0000000
#> 
#> $CQ
#>                  PLANTS  BACTERIA DETRITUS FEEDERS CARNIVORES  DETRITUS
#> PLANTS                0 0.0000000         0.000000  0.0000000 0.0000000
#> BACTERIA              1 0.0000000         0.000000  0.1865961 0.6903851
#> DETRITUS FEEDERS      1 0.5806865         0.000000  0.0000000 0.8460570
#> CARNIVORES            1 0.0000000         0.891124  0.0000000 0.5486486
#> DETRITUS              1 0.0000000         0.000000  0.0000000 0.0000000
#> 
#> $CD
#>                         PLANTS      BACTERIA DETRITUS FEEDERS    CARNIVORES
#> PLANTS            0.000000e+00  8.344691e-05     8.344691e-05  8.344691e-05
#> BACTERIA         -8.344691e-05  0.000000e+00     2.240311e-05 -8.850401e-06
#> DETRITUS FEEDERS -8.344691e-05 -2.240311e-05     0.000000e+00  3.882096e-04
#> CARNIVORES       -8.344691e-05  8.850401e-06    -3.882096e-04  0.000000e+00
#> DETRITUS         -8.344691e-05  7.254992e-05     8.890890e-05  5.765539e-05
#>                       DETRITUS
#> PLANTS            8.344691e-05
#> BACTERIA         -7.254992e-05
#> DETRITUS FEEDERS -8.890890e-05
#> CARNIVORES       -5.765539e-05
#> DETRITUS          0.000000e+00
#> 
#> $CR
#>                  PLANTS   BACTERIA DETRITUS FEEDERS CARNIVORES   DETRITUS
#> PLANTS                0  1.0000000        1.0000000  1.0000000  1.0000000
#> BACTERIA             -1  0.0000000        0.5806865 -0.1865961 -0.6903851
#> DETRITUS FEEDERS     -1 -0.5806865        0.0000000  0.8911240 -0.8460570
#> CARNIVORES           -1  0.1865961       -0.8911240  0.0000000 -0.5486486
#> DETRITUS             -1  0.6903851        0.8460570  0.5486486  0.0000000
#> 
#> $CA
#>      PLANTS  BACTERIA DETRITUS FEEDERS CARNIVORES  DETRITUS
#> [1,]      0 0.2677194        0.1226211 0.01903096 0.5906285
#> [2,]      0 0.0000000        1.0000000 0.00000000 0.0000000
#> [3,]      0 0.0000000        0.0000000 1.00000000 0.0000000
#> [4,]      0 1.0000000        0.0000000 0.00000000 0.0000000
#> [5,]      0 0.6181272        0.3469538 0.03491900 0.0000000
#> 
#> $CDep
#>      PLANTS  BACTERIA DETRITUS FEEDERS CARNIVORES DETRITUS
#> [1,]      0 0.0000000        0.0000000  0.0000000        0
#> [2,]      0 0.0000000        0.0622461  0.0000000        0
#> [3,]      0 0.0000000        0.0000000  0.3093601        0
#> [4,]      0 0.2352179        0.0000000  0.0000000        0
#> [5,]      0 0.3940869        0.2212004  0.0222626        0
#> 
#> $sc
#>           PLANTS         BACTERIA DETRITUS FEEDERS       CARNIVORES 
#>     0.0003337877    -0.0001424441     0.0001934507    -0.0005204615 
#>         DETRITUS 
#>     0.0001356673 
#> 
#> $psc
#>           PLANTS         BACTERIA DETRITUS FEEDERS       CARNIVORES 
#>         50.35221        -21.48784         29.18223        -78.51216 
#>         DETRITUS 
#>         20.46555 
#> 
#> $ns
#>          TSC 
#> 0.0006629056 
#>