Ecological Network Environs

Calculates the environs for an ecological network.

Usage

enaEnviron(
  x,
  input = TRUE,
  output = TRUE,
  type = "unit",
  err.tol = 1e-10,
  balance.override = FALSE
)

Arguments

x

A network object.

input

Should the input environ be calculated?

output

Should the output environ be calculated?

type

Specifies the type of environs ("unit" or "realized") to be calculated.

err.tol

Error threshold for numerical error fluctuations in flows. Values below err.tol will be set to zero.

balance.override

Logical specifying whether (TRUE) or not (FALSE) the model needs to be balanced prior to calculations. If TRUE and the model is not balanced, environs will not be calculated.

Value

The function returns the input, output or both environs depending upon which were requested.

References

Fath, B.D. and S.R. Borrett. 2006. A MATLAB function for network environ analysis. Environmental Modelling & Software 21:375-405.

Author

Stuart R. Borrett Matthew K. Lau

Examples

data(troModels)
enaEnviron(troModels[[6]])
#> $input
#> $input$PLANTS
#>                  PLANTS BACTERIA DETRITUS FEEDERS CARNIVORES DETRITUS y
#> PLANTS               -1        0                0          0        0 1
#> BACTERIA              0        0                0          0        0 0
#> DETRITUS FEEDERS      0        0                0          0        0 0
#> CARNIVORES            0        0                0          0        0 0
#> DETRITUS              0        0                0          0        0 0
#> z                     1        0                0          0        0 0
#> 
#> $input$BACTERIA
#>                      PLANTS  BACTERIA DETRITUS FEEDERS  CARNIVORES    DETRITUS
#> PLANTS           -0.9332703  0.000000      0.000000000  0.00000000  0.93327028
#> BACTERIA          0.0000000 -1.169351      0.001213296  0.00000000  0.16813787
#> DETRITUS FEEDERS  0.0000000  0.000000     -0.038566625  0.01754939  0.02101723
#> CARNIVORES        0.0000000  0.000000      0.000000000 -0.01754939  0.01754939
#> DETRITUS          0.0000000  1.169351      0.037353329  0.00000000 -1.20670450
#> z                 0.9332703  0.000000      0.000000000  0.00000000  0.06672972
#>                  y
#> PLANTS           0
#> BACTERIA         1
#> DETRITUS FEEDERS 0
#> CARNIVORES       0
#> DETRITUS         0
#> z                0
#> 
#> $input$`DETRITUS FEEDERS`
#>                      PLANTS   BACTERIA DETRITUS FEEDERS  CARNIVORES    DETRITUS
#> PLANTS           -0.9332703  0.0000000       0.00000000  0.00000000  0.93327028
#> BACTERIA          0.0000000 -0.2008109       0.03267303  0.00000000  0.16813787
#> DETRITUS FEEDERS  0.0000000  0.0000000      -1.03856662  0.01754939  0.02101723
#> CARNIVORES        0.0000000  0.0000000       0.00000000 -0.01754939  0.01754939
#> DETRITUS          0.0000000  0.2008109       1.00589360  0.00000000 -1.20670450
#> z                 0.9332703  0.0000000       0.00000000  0.00000000  0.06672972
#>                  y
#> PLANTS           0
#> BACTERIA         0
#> DETRITUS FEEDERS 1
#> CARNIVORES       0
#> DETRITUS         0
#> z                0
#> 
#> $input$CARNIVORES
#>                      PLANTS   BACTERIA DETRITUS FEEDERS CARNIVORES    DETRITUS
#> PLANTS           -0.9332703  0.0000000       0.00000000   0.000000  0.93327028
#> BACTERIA          0.0000000 -0.2008109       0.03267303   0.000000  0.16813787
#> DETRITUS FEEDERS  0.0000000  0.0000000      -1.03856662   1.017549  0.02101723
#> CARNIVORES        0.0000000  0.0000000       0.00000000  -1.017549  0.01754939
#> DETRITUS          0.0000000  0.2008109       1.00589360   0.000000 -1.20670450
#> z                 0.9332703  0.0000000       0.00000000   0.000000  0.06672972
#>                  y
#> PLANTS           0
#> BACTERIA         0
#> DETRITUS FEEDERS 0
#> CARNIVORES       1
#> DETRITUS         0
#> z                0
#> 
#> $input$DETRITUS
#>                      PLANTS   BACTERIA DETRITUS FEEDERS  CARNIVORES    DETRITUS
#> PLANTS           -0.9332703  0.0000000      0.000000000  0.00000000  0.93327028
#> BACTERIA          0.0000000 -0.1693512      0.001213296  0.00000000  0.16813787
#> DETRITUS FEEDERS  0.0000000  0.0000000     -0.038566625  0.01754939  0.02101723
#> CARNIVORES        0.0000000  0.0000000      0.000000000 -0.01754939  0.01754939
#> DETRITUS          0.0000000  0.1693512      0.037353329  0.00000000 -1.20670450
#> z                 0.9332703  0.0000000      0.000000000  0.00000000  0.06672972
#>                  y
#> PLANTS           0
#> BACTERIA         0
#> DETRITUS FEEDERS 0
#> CARNIVORES       0
#> DETRITUS         1
#> z                0
#> 
#> 
#> $output
#> $output$PLANTS
#>                  PLANTS   BACTERIA DETRITUS FEEDERS  CARNIVORES    DETRITUS
#> PLANTS               -1  0.0000000      0.000000000  0.00000000  0.79408083
#> BACTERIA              0 -0.4343412      0.006258519  0.00000000  0.13351506
#> DETRITUS FEEDERS      0  0.0000000     -0.198937442  0.03087536  0.01668938
#> CARNIVORES            0  0.0000000      0.000000000 -0.03087536  0.01393563
#> DETRITUS              0  0.4343412      0.192678923  0.00000000 -0.95822091
#> z                     1  0.0000000      0.000000000  0.00000000  0.00000000
#>                           y
#> PLANTS           0.20591917
#> BACTERIA         0.29456761
#> DETRITUS FEEDERS 0.15137270
#> CARNIVORES       0.01693972
#> DETRITUS         0.33120080
#> z                0.00000000
#> 
#> $output$BACTERIA
#>                  PLANTS   BACTERIA DETRITUS FEEDERS  CARNIVORES     DETRITUS
#> PLANTS                0  0.0000000       0.00000000  0.00000000  0.000000000
#> BACTERIA              0 -1.1693512       0.01684944  0.00000000  0.359454730
#> DETRITUS FEEDERS      0  0.0000000      -0.09197564  0.01427474  0.007716077
#> CARNIVORES            0  0.0000000       0.00000000 -0.01427474  0.006442924
#> DETRITUS              0  0.1693512       0.07512620  0.00000000 -0.373613731
#> z                     0  1.0000000       0.00000000  0.00000000  0.000000000
#>                            y
#> PLANTS           0.000000000
#> BACTERIA         0.793046998
#> DETRITUS FEEDERS 0.069984817
#> CARNIVORES       0.007831818
#> DETRITUS         0.129136367
#> z                0.000000000
#> 
#> $output$`DETRITUS FEEDERS`
#>                  PLANTS    BACTERIA DETRITUS FEEDERS CARNIVORES    DETRITUS
#> PLANTS                0  0.00000000      0.000000000  0.0000000  0.00000000
#> BACTERIA              0 -0.08420272      0.001213296  0.0000000  0.02588364
#> DETRITUS FEEDERS      0  0.00000000     -1.038566625  0.1611869  0.08712807
#> CARNIVORES            0  0.00000000      0.000000000 -0.1611869  0.07275194
#> DETRITUS              0  0.08420272      0.037353329  0.0000000 -0.18576365
#> z                     0  0.00000000      1.000000000  0.0000000  0.00000000
#>                           y
#> PLANTS           0.00000000
#> BACTERIA         0.05710578
#> DETRITUS FEEDERS 0.79025162
#> CARNIVORES       0.08843499
#> DETRITUS         0.06420761
#> z                0.00000000
#> 
#> $output$CARNIVORES
#>                  PLANTS   BACTERIA DETRITUS FEEDERS  CARNIVORES     DETRITUS
#> PLANTS                0  0.0000000      0.000000000  0.00000000  0.000000000
#> BACTERIA              0 -0.2468772      0.003557309  0.00000000  0.075889256
#> DETRITUS FEEDERS      0  0.0000000     -0.113074992  0.01754939  0.009486157
#> CARNIVORES            0  0.0000000      0.000000000 -1.01754939  0.459272292
#> DETRITUS              0  0.2468772      0.109517683  0.00000000 -0.544647706
#> z                     0  0.0000000      0.000000000  1.00000000  0.000000000
#>                           y
#> PLANTS           0.00000000
#> BACTERIA         0.16743067
#> DETRITUS FEEDERS 0.08603944
#> CARNIVORES       0.55827710
#> DETRITUS         0.18825279
#> z                0.00000000
#> 
#> $output$DETRITUS
#>                  PLANTS   BACTERIA DETRITUS FEEDERS  CARNIVORES    DETRITUS
#> PLANTS                0  0.0000000      0.000000000  0.00000000  0.00000000
#> BACTERIA              0 -0.5469735      0.007881463  0.00000000  0.16813787
#> DETRITUS FEEDERS      0  0.0000000     -0.250525431  0.03888188  0.02101723
#> CARNIVORES            0  0.0000000      0.000000000 -0.03888188  0.01754939
#> DETRITUS              0  0.5469735      0.242643968  0.00000000 -1.20670450
#> z                     0  0.0000000      0.000000000  0.00000000  1.00000000
#>                           y
#> PLANTS           0.00000000
#> BACTERIA         0.37095418
#> DETRITUS FEEDERS 0.19062631
#> CARNIVORES       0.02133249
#> DETRITUS         0.41708701
#> z                0.00000000
#> 
#>