Storage Analyses of Ecological Networks

Calculates storage-based Ecological Network Analyses.

Usage

enaStorage(x, balance.override = FALSE)

Arguments

x

A network object. This This includes all weighted flows into and out of each vertex as well as the amount of energy–matter stored at each vertex.

balance.override

LOGICAL: should an imbalanced model be analyzed? If FALSE, the functions checks to make sure the network model provided is at steady-state. If TRUE, then the function will run without ensuring that the model meets the steady-state assumption.

Value

X

The storage values themselves.

C

output or donor-storage normalized output-oriented direct flow intensity matrix (Jacobian community matrix)

S

dimensionalized integral output community matrix

VS

variance in expected output-oriented residance times (Barber 1979)

Q

integral output storage matrix - non-dimensional

CP

input or recipient-storage normalized oriented flow intensity matrix (Jacobian community matrix)

SP

dimensionalized integral input community matrix

VSP

variance in expected input-oriented residance times (Barber 1979)

QP

integral input storage matrix - non-dimensional

dt

selected time step to create P, PP, Q and QP - smallest whole number to make diag(C) nonnegative

RT

node residence time (storage/outputs), units of inverse time

ns

vector of the storage based whole system network statistics. These statistics include total system storage (TSS), storage cycling index (CIS), Boundary storage intensity (BSI), Direct storage intensity (DSI), Indirect storage intensity (ISI), realized ratio of indirect-to-direct storage (ID.S), unit input-oriented ratio of indirect-to-direct storage intensities (IDS.I), unit output ratio of indirect-to-direct storage intensities (IDS.O), input-oriented storage-based network homogenization (HMG.S.I), output-oriented storage-based network homogenization (HMG.S.O), input-oriented storage-based network amplification (AMP.S.I), output-oriented storage-based network amplification (AMP.S.O), Storage from Boundary flow (mode0.S), storage from internal first passage flow (mode1.S), storage from cycled flow (mode2.S), dissipative equivalent to mode1.S (mode3.S), dissipative equivalent to mode0.S (mode4.S), Average Residence Time (ART)

References

Barber, M. C. 1978a. A Markovian Model for Ecosystem Flow Analysis. Ecol. Model. 5(3):193-206.

Barber, M. C. 1978b. A Retrospective Markovian Model for Ecosystem Resource Flow. Ecol. Model. 5(2): 125-35.

Barber, M. C. 1979. A Note Concerning Time Parameterization of Markovian Models of Ecosystem Flow Analysis. Ecol. Model. 6(4): 323-28.

Matis, J. H., Patten, B. C. 1981. Environ analysis of linear compartmental systems: the static, time invariant case. Bulletin of the International Statistical Institute, 48: 527-565.

Fath, B. D., Patten, B. C. 1999. Review of the foundations of network enviorn analysis. Ecosystems 2:167-179.

Fath, B. D. Patten, B. C., Choi, J. 2001. Compementarity of ecological goal functions. Journal of Theoretical Biology 208: 493-506.

Fath, B. D., Borrett, S. R. 2006. A MATLAB function for Network Environ Analysis. Environmental Modelling & Software 21:375-405.

See also

Author

Matthew K. Lau Stuart R. Borrett

Examples

data(oyster)
S <- enaStorage(oyster)
attributes(S)
#> $names
#>  [1] "X"   "C"   "P"   "S"   "VS"  "Q"   "CP"  "PP"  "SP"  "VSP" "QP"  "dt" 
#> [13] "RT"  "ns" 
#>