Repeated Application the Balance Function

This function repeatedly balances a model, sequentially with the output being passed back to the balance function, until it is within tolerance or the maximum number of iterations is reached.

Usage

force.balance(x, tol = 5, max.itr = 10, method = "AVG2")

Arguments

x

A network object.

tol

Percent error tolerance for difference between inputs and outputs.

max.itr

Maximum number iterations.

method

The balancing method to use, see balance. DEFAULT = AVG2.

Value

Returns a balanced network model.

References

Allesina, S., Bondavalli, C., 2003.Steady state of ecosystem flow networks: a comparison between balancing procedures.Ecological Modelling 165(2-3):231-239.

See also

Author

Matthew K. Lau Stuart R. Borrett

Examples

data(troModels)
ssCheck(troModels[[1]])
#> [1] TRUE
fb.model = force.balance(troModels[[2]])  #produces a balanced model