Calculate the exponent of a given matrix

Function for calculating the pathway proliferation of flows in a network model through matrix exponentiation.

Usage

mExp(x = "matrix", n = 2)

Arguments

x

A matrix.

n

Desired exponent (i.e. the path length).

Value

Returns an exponentiated flow matrix.

Details

This function was originally designed by Alberto Monteiro in the following R help thread: https://stat.ethz.ch/pipermail/r-help/2007-May/131330.html.

See also

Author

Alberto Monteiro (https://stat.ethz.ch/pipermail/r-help/2007-May/131330.html) Matthew K. Lau