Relational change compared between two matrices.

Identifies the signs and pairwise relationsips of two matrices and compares the difference between them.

Usage

relationalChange(x = "Direct.U", y = "Integral.U")

Arguments

x

x is a square matrix of real numbers. While this function is more general, the initial intention was for this to be the direct utility matrix.

y

y is a square matrix of real numbers. While this function is more general, the initial intention was for this to be the integral utility matrix or the mixed trophic impacts matrix.

Value

Direct.Signs

A sign matrix for matrix x.

Integral.Signs

A sign matrix for matrix x.

Direct.Relations

A matrix of the pairwise sign relationships for matrix x.

Integral.Relations

A matrix of the pairwise signed relationships in matrix y.

Relations.Table

A table that summarizes the relations.

Changed.Table

A summary table of only the pariwise relationships that changed between x and y.

ns

A vector of network statisitcs which currently includes one whole-network statistic - a ratio of the relationships changed between x and y.

Note

This function is called by enaUtility and enaMTI to summarize results.

See also

Author

Stuart R. Borrett