Globally reorients matrices

Changes the orientation of output matrices.

Usage

set.orient(x = c("rc", "school"))

Arguments

x

The school determining the orientation

Details

The enaR package as a whole, and the broader network analysis community, assumes a row to column orientation; thus, the default orientation for the package is row to column (DEFAULT = 'rc'). However, functions from the Patten school were orignially developed to conduct calculations and produce output in the column to row orientation. In order to facilitate the use of these functions, we also provide the option for users to return output in the orientation of the "school" (i.e. Patten results will be column to row oriented) by setting the global orientation to "school" using this fuction. Orientation setting. If "rc" (DEFAULT), all matrix output will be returned in row (=input) to column (=output) orientation, regardless of school. If "school", then output matrices from functions from particular ENA schools will be oriented as expected in that school (i.e. Patten = column-row or Ulanowicz = row-column). Note, that all functions in the enaR package expect input matrices to be oriented row-column.

See also

Author

Matthew K. Lau Stuart R. Borrett

Examples

original.orientation = get.orient()
original.orientation
#> [1] "rc"
set.orient("school")
#> Warning: NOTE: output of functions from a particular analytical school will be returned in the standard orientation of that school.
get.orient()
#> [1] "school"
set.orient("rc")
get.orient()
#> [1] "rc"
set.orient(original.orientation)