Signs and summary of input matrix

Identifies the signs and pairwise relationsips of a given matrix. This includes also returns a summary table that provides the ecological name of each pairwise realtionship, and a summary of the counts.

Usage

signs(x = "matrix")

Arguments

x

a square matrix of real numbers. While this function is more general, the initail intention was for this to be a utility matrix or the mixed trophic impacts matrix.

Value

sign

A sign matrix for matrix x.

relations

A matrix of the pairwise signed relationships in x.

rs.tab

Table summarizing the pairwise relationships and identifying their ecological label.

relationship.counts

A count of the different kinds of pairwise relationships found in matrix x.

Note

This function is called by relationalChange, and was created to generate more informative output from enaUtility and enaMTI.

See also

Author

Stuart R. Borrett