Checks if the given network is out of balance by a given tolerance threshold

This function supports the balancing process by checking if the inputs and outputs of a given network model are within acceptable limits.

Usage

ssCheck(x, tol = 5, more = FALSE, zero.na = TRUE)

Arguments

x

A network object.

tol

The threshold for balance in percent difference between input and outputs.

more

LOGICAL: should more detailed results be returned?

zero.na

LOGICAL: should NA values be changed to zeros?

Value

Returns a logical value stating if the model is within acceptable limits of balance (TRUE) or if it is not (FALSE).

References

Fath, B.D. and S.R. Borrett. 2006. A MATLAB function for network environ analysis. Environmental Modelling & Software 21:375-405.

See also

Author

Matthew K. Lau Stuart R. Borrett

Examples

data(troModels)
ssCheck(troModels[[2]])
#> [1] TRUE
ssCheck(troModels[[6]])
#> [1] TRUE