structural statistics

This function returns several network statistics that describe a network.

Usage

structure.statistics(A = "adjacency matrix")

Arguments

A

An adjacency matrix.

Value

n

Number of nodes in A.

L

Number of direct connections in A.

C

Connectivity of A.

LD

Link density.

lam1A

First dominant eigenvalue of A.

mlam1A

Multiplicity of the dominant eigenvalue.

lam2A

Magnitude of the second largest eigenvalue.

rho

Damping ratio (see Caswell 2001).

R

Distance of lam1A from the bulk of the eigen spectrum.

d

Difference between the dominant eigenvalue and the link density.

no.scc

Number of strongly connected components.

no.scc.big

Number of strongly connected components greater than 1.

pscc

Precent of nodes participating in a strongly connected component.

References

Fath, B. D., Borrett, S. R. 2006. A Matlab function for Network Environ Analysis. Environ. Model. Softw. 21, 375-405.

See also

Author

Matthew K. Lau Stuart R. Borrett