Extracts network object into a list

Separates the components of a network object into separate components within a list. This includes inputs, exports, respirations, outputs (exports + respirations), storage, and internal flows.

Usage

unpack(x = "network object")

Arguments

x

A network object. This includes all weighted flows into and out of each node.

Value

F

matrix of flows from each node to each node oreinted row to column.

z

Node boundary inputs.

r

Node boundary loss from respiration.

e

Node boundary loss due to exportation

y

Node boundary loss; summation of r and e

X

Node storage or biomass

living

Logical vector indicating whether each node is living or not

Details

Flows are oriented from row to column.

See also

Author

Matthew K. Lau Stuart R. Borrett

Examples

data(troModels)
unpack(troModels[[6]])
#> $F
#>                  PLANTS BACTERIA DETRITUS FEEDERS CARNIVORES DETRITUS
#> PLANTS                0        0                0          0     8881
#> BACTERIA              0        0               75          0     1600
#> DETRITUS FEEDERS      0        0                0        370      200
#> CARNIVORES            0        0                0          0      167
#> DETRITUS              0     5205             2309          0        0
#> 
#> $z
#> [1] 11184     0     0     0   635
#> 
#> $r
#> [1] 2003 3275 1814  203 3109
#> 
#> $e
#> [1] 300 255   0   0 860
#> 
#> $y
#> [1] 2303 3530 1814  203 3969
#> 
#> $X
#> [1]  285.0  117.0   60.0   17.0 3579.4
#> 
#> $living
#> [1]  TRUE  TRUE  TRUE  TRUE FALSE
#>