Write enaR models to an EcoNet formatted file.

Creates an EcoNet model from an enaR network object that can be used with the online interface for EcoNet.

Usage

write.EcoNet(
  x = "model",
  file = "file path",
  mn = "ena_model",
  zero.flows = FALSE
)

Arguments

x

Network object.

file

The file name or path. If a simple file name is given, this function uses the current working directory by default.

mn

The model name that EcoNet will use. The DEFAULT is 'ena_model'.

zero.flows

LOGICAL: should zero flow values be written?

Value

An EcoNet formatted text file is created from the model, which can be input at http://eco.engr.uga.edu.

References

About EcoNet (http://eco.engr.uga.edu/DOC/econet1.html) Kazanci, C. 2009. Handbook of Ecological Modelling and Informatics, by WIT Press.

Author

Matthew K. Lau